Minor changes
Former-commit-id: db0a9ae96f7c39071afa4773524a01411a30a66d

+++ renderer/scenes/blinn_scenes.c
@@ -279,8 +279,8 @@ scene_t *blinn_craftsman_scene(void) {
     int num_sparks = ARRAY_SIZE(spark_transforms);
     int i;
 
-    assert((int)ARRAY_SIZE(materials) <= num_meshes);
-    assert((int)ARRAY_SIZE(transforms) == num_meshes);
+    assert(ARRAY_SIZE(materials) <= num_meshes);
+    assert(ARRAY_SIZE(transforms) == num_meshes);
 
     translation = mat4_translate(-1.668f, -27.061f, -10.834f);
     scale = mat4_scale(0.016f, 0.016f, 0.016f);
@@ -288,7 +288,7 @@ scene_t *blinn_craftsman_scene(void) {
     for (i = 0; i < num_meshes; i++) {
         mat4_t transform = mat4_mul_mat4(root, transforms[i]);
         blinn_material_t material;
-        if (i < (int)ARRAY_SIZE(materials)) {
+        if (i < ARRAY_SIZE(materials)) {
             material = materials[i];
         } else {
             material = materials[ARRAY_SIZE(materials) - 1];
